+++ Debug/Core/Src/gpio.cyclo
@@ -1,0 +1,1 @@
+../Core/Src/gpio.c:42:6:MX_GPIO_Init	1
