+++ Debug/Core/Src/gpio.su
@@ -1,0 +1,1 @@
+../Core/Src/gpio.c:42:6:MX_GPIO_Init	48	static
