+++ Debug/Core/Src/main.cyclo
@@ -1,5 +1,4 @@
-../Core/Src/main.c:67:5:main	9
-../Core/Src/main.c:160:6:SystemClock_Config	3
-../Core/Src/main.c:207:13:MX_LPUART1_UART_Init	5
-../Core/Src/main.c:254:13:MX_GPIO_Init	1
-../Core/Src/main.c:290:6:Error_Handler	1
+../Core/Src/main.c:65:5:main	9
+../Core/Src/main.c:159:6:SystemClock_Config	3
+../Core/Src/main.c:202:6:EXTI15_10_IRQHandler	1
+../Core/Src/main.c:220:6:Error_Handler	1
