+++ Debug/Core/Src/main.su
@@ -1,5 +1,4 @@
-../Core/Src/main.c:67:5:main	144	static
-../Core/Src/main.c:160:6:SystemClock_Config	88	static
-../Core/Src/main.c:207:13:MX_LPUART1_UART_Init	8	static
-../Core/Src/main.c:254:13:MX_GPIO_Init	48	static
-../Core/Src/main.c:290:6:Error_Handler	4	static,ignoring_inline_asm
+../Core/Src/main.c:65:5:main	144	static
+../Core/Src/main.c:159:6:SystemClock_Config	88	static
+../Core/Src/main.c:202:6:EXTI15_10_IRQHandler	8	static
+../Core/Src/main.c:220:6:Error_Handler	4	static,ignoring_inline_asm
