+++ Debug/Core/Src/stm32g4xx_hal_msp.cyclo
@@ -1,3 +1,1 @@
 ../Core/Src/stm32g4xx_hal_msp.c:63:6:HAL_MspInit	1
-../Core/Src/stm32g4xx_hal_msp.c:90:6:HAL_UART_MspInit	3
-../Core/Src/stm32g4xx_hal_msp.c:138:6:HAL_UART_MspDeInit	2
