+++ Debug/Core/Src/stm32g4xx_hal_msp.su
@@ -1,3 +1,1 @@
 ../Core/Src/stm32g4xx_hal_msp.c:63:6:HAL_MspInit	16	static
-../Core/Src/stm32g4xx_hal_msp.c:90:6:HAL_UART_MspInit	112	static
-../Core/Src/stm32g4xx_hal_msp.c:138:6:HAL_UART_MspDeInit	16	static
