+++ Debug/Core/Src/stm32g4xx_it.cyclo
@@ -7,4 +7,3 @@
 ../Core/Src/stm32g4xx_it.c:157:6:DebugMon_Handler	1
 ../Core/Src/stm32g4xx_it.c:170:6:PendSV_Handler	1
 ../Core/Src/stm32g4xx_it.c:183:6:SysTick_Handler	1
-../Core/Src/stm32g4xx_it.c:204:6:EXTI15_10_IRQHandler	1
