+++ Debug/Core/Src/stm32g4xx_it.su
@@ -7,4 +7,3 @@
 ../Core/Src/stm32g4xx_it.c:157:6:DebugMon_Handler	4	static
 ../Core/Src/stm32g4xx_it.c:170:6:PendSV_Handler	4	static
 ../Core/Src/stm32g4xx_it.c:183:6:SysTick_Handler	8	static
-../Core/Src/stm32g4xx_it.c:204:6:EXTI15_10_IRQHandler	8	static
