+++ Debug/Core/Src/usart.cyclo
@@ -1,0 +1,3 @@
+../Core/Src/usart.c:31:6:MX_LPUART1_UART_Init	5
+../Core/Src/usart.c:73:6:HAL_UART_MspInit	3
+../Core/Src/usart.c:114:6:HAL_UART_MspDeInit	2
