+++ Debug/Core/Src/usart.su
@@ -1,0 +1,3 @@
+../Core/Src/usart.c:31:6:MX_LPUART1_UART_Init	8	static
+../Core/Src/usart.c:73:6:HAL_UART_MspInit	112	static
+../Core/Src/usart.c:114:6:HAL_UART_MspDeInit	16	static
