+++ Debug/objects.list
@@ -1,9 +1,11 @@
+"./Core/Src/gpio.o"
 "./Core/Src/main.o"
 "./Core/Src/stm32g4xx_hal_msp.o"
 "./Core/Src/stm32g4xx_it.o"
 "./Core/Src/syscalls.o"
 "./Core/Src/sysmem.o"
 "./Core/Src/system_stm32g4xx.o"
+"./Core/Src/usart.o"
 "./Core/Startup/startup_stm32g431rbtx.o"
 "./Drivers/BSP/STM32G4xx_Nucleo/stm32g4xx_nucleo.o"
 "./Drivers/STM32G4xx_HAL_Driver/Src/stm32g4xx_hal.o"
