+++ LED_OnOff_Command.ioc
@@ -38,7 +38,7 @@ Mcu.UserConstants=
 Mcu.UserName=STM32G431RBTx
 MxCube.Version=6.14.1
 MxDb.Version=DB.6.0.141
-NUCLEO-G431RB.BUTTON=1
+NUCLEO-G431RB.BUTTON=2
 NUCLEO-G431RB.IPParameters=LD1,BUTTON
 NUCLEO-G431RB.LD1=true
 NVIC.BusFault_IRQn=true\:0\:0\:false\:false\:true\:false\:false\:false
@@ -77,6 +77,9 @@ PB3.GPIOParameters=GPIO_Label
 PB3.GPIO_Label=T_SWO
 PB3.Locked=true
 PB3.Signal=SYS_JTDO-SWO
+PC13.GPIOParameters=GPIO_Label,GPIO_ModeDefaultEXTI
+PC13.GPIO_Label=Button
+PC13.GPIO_ModeDefaultEXTI=GPIO_MODE_IT_FALLING
 PC13.Locked=true
 PC13.Signal=GPXTI13
 PC14-OSC32_IN.GPIOParameters=GPIO_Label
@@ -101,7 +104,7 @@ ProjectManager.BackupPrevious=false
 ProjectManager.CompilerLinker=GCC
 ProjectManager.CompilerOptimize=6
 ProjectManager.ComputerToolchain=false
-ProjectManager.CoupleFile=false
+ProjectManager.CoupleFile=true
 ProjectManager.CustomerFirmwarePackage=
 ProjectManager.DefaultFWLocation=true
 ProjectManager.DeletePrevious=true
@@ -183,3 +186,4 @@ VP_SYS_VS_Systick.Mode=SysTick
 VP_SYS_VS_Systick.Signal=SYS_VS_Systick
 board=NUCLEO-G431RB
 boardIOC=true
+isbadioc=false
